\name{webpick}
\alias{webpick}

\title{
enviPick web browser UI
}

\description{
Run enviPick conveniently from a web browser-based user interface
}

\usage{
  webpick()
}

\arguments{  
  check webpage
}


\section{Warnings}{May not work with Microsoft Internet Explorer; better choose a different default browser (e.g., Google Chrome).}


\author{Martin  Loos}








